Separate loading configuration and packages
This very messy, and it needs further work before merging. One issue resulted
from the separation is duplication of code, we see the two loops and two
times the module_d gets defined. I'm sure you would fine many more

As of this commit, everything works fine, I've tested multiple time.

One advantage to this might be the ability to call load_packages from an
active session and have the plugin updated or install with the addition
of minpac.update to it.